+++ tests/gtest/aviftest_helpers.h
@@ -22,7 +22,7 @@ using AvifDecoderPtr =
 
 class AvifRwData : public avifRWData {
  public:
-  AvifRwData() : avifRWData({}) {}
+  AvifRwData() : avifRWData{nullptr, 0} {}
   ~AvifRwData() { avifRWDataFree(this); }
 };
 
@@ -54,7 +54,10 @@ void FillImageChannel(avifRGBImage* image, uint32_t channel_offset,
                       uint32_t value);
 
 // Returns true if both images have the same features and pixel values.
-bool AreImagesEqual(const avifImage& image1, const avifImage& image2);
+// If ignore_alpha is true, the alpha channel is not taken into account in the
+// comparison.
+bool AreImagesEqual(const avifImage& image1, const avifImage& image2,
+                    bool ignore_alpha = false);
 
 //------------------------------------------------------------------------------